requirements.txt
# Data & Analysis
pandas==2.1.4
numpy==1.26.2
yfinance==0.2.33
scipy==1.11.4

# Web Scraping
beautifulsoup4==4.12.2
requests==2.31.0
selenium==4.16.0
lxml==4.9.4

# Database
sqlalchemy==2.0.23

# Backtesting & Reports
matplotlib==3.8.2
seaborn==0.13.0
schedule==1.2.1

# Notifications
python-telegram-bot==20.7

# Scheduling & Utilities
APScheduler==3.10.4
python-dotenv==1.0.0
pytz==2023.3
loguru==0.7.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
